Allow build parameters in branch configuration
Perform build parameter substitution in branch names.  The user can now set
the branch to build as ${branch} and then pass in a parameter at build time
telling Hudson which version to build.

Previously "origin/" was prepended to all branch names.  This is no longer
done.  This allows you to pass in anything that is resolvable to a commit,
including a commit itself, but if you want to follow a branch you need to
explicitly state "origin/[branch-name]".

The hudson version needed to be upgraded in order to have access to this
new functionality.  Removing the Eclipse project files as they can be
generated using mvn eclipse:eclipse (or mvn eclipse:m2eclipse if you use
the m2eclipse plugin).