Improve display of source code changes
Changes made to the gitweb configuration were not being saved.  Fixed that
issue, but the fix may cause issues with other RepositoryBrowsers because
it essentially hardcodes gitweb.

Fixed the message formatting in other places.  Only show the title and not
the whole message body in the change list.  In order to see the whole
message you need to click on details.  The details page wraps the entire
message in a pre tag so that it will be formatted correctly.